Add very basic ability to "Join" a project.
By default, people join at the "INTERESTED" level and are only visible to leaders and admins on the project page.

If a project has leaders or involved members, they'll be shown publically on the project page.

Note: once a user joins, their membership type is maintained even if they unjoin. This way if they change their mind, they don't have to be re-approved.